requirements.txt
pandas

apache-airflow-providers-postgres

mysql-connector-python

apache-airflow-providers-mysql

python-dotenv

# Dashboard UI
streamlit>=1.38.0
plotly>=5.24.0

# Kafka Integration
kafka-python>=2.0.2
confluent-kafka>=2.3.0

# Database connectors
pymysql>=1.1.0
sqlalchemy>=1.4.0

# Data processing
numpy>=1.24.0
pytz>=2023.3

# Monitoring and logging
python-json-logger>=2.0.7